+++ packages/g/src/__tests__/canvas.spec.ts
@@ -6,8 +6,9 @@ import sinon from 'sinon';
 // @ts-ignore
 import sinonChai from 'sinon-chai';
 
-import { Group, Circle, Canvas, Text, Rect, ElementEvent } from '@antv/g';
-import { Renderer as CanvasRenderer } from '@antv/g-canvas';
+import { Group, Circle, Canvas, Text, Rect, ElementEvent } from '../';
+import { Renderer as CanvasRenderer } from '../../../g-canvas';
+import type { FederatedPointerEvent } from '../dom';
 
 chai.use(chaiAlmost(0.0001));
 chai.use(sinonChai);
@@ -177,7 +178,7 @@ describe('Canvas', () => {
 
     canvas.addEventListener(
       'pointerdown',
-      (e) => {
+      (e: FederatedPointerEvent) => {
         // coordinates
         expect(e.clientX).to.be.eqls(100);
         expect(e.clientY).to.be.eqls(100);
@@ -218,7 +219,7 @@ describe('Canvas', () => {
 
     canvas.addEventListener(
       'pointerdown',
-      (e) => {
+      (e: FederatedPointerEvent) => {
         // coordinates
         expect(e.clientX).to.be.eqls(100);
         expect(e.clientY).to.be.eqls(100);