+++ packages/g/src/__tests__/event.spec.ts
@@ -6,9 +6,9 @@ import sinon from 'sinon';
 // @ts-ignore
 import sinonChai from 'sinon-chai';
 // @ts-ignore
-import { Group, Circle, Canvas, Text, Rect, ElementEvent } from '@antv/g';
-import { Renderer as CanvasRenderer } from '@antv/g-canvas';
-import { Plugin } from '@antv/g-plugin-css-select';
+import { Group, Circle, Canvas, Text, Rect, ElementEvent } from '../';
+import { Renderer as CanvasRenderer } from '../../../g-canvas';
+import { Plugin } from '../../../g-plugin-css-select';
 import interact from 'interactjs';
 
 chai.use(chaiAlmost());