+++ packages/g/src/display-objects/mixins/index.ts
@@ -1,3 +1,0 @@
-export * from './Visible';
-export * from './Animatable';
-export * from './Transformable';