+++ packages/g/src/dom/__tests__/animation.timeline.spec.ts
@@ -6,8 +6,8 @@ import sinon from 'sinon';
 // @ts-ignore
 import sinonChai from 'sinon-chai';
 
-import { Group, Circle, Canvas, Text, Rect, ElementEvent } from '@antv/g';
-import { Renderer as CanvasRenderer } from '@antv/g-canvas';
+import { Group, Circle, Canvas, Text, Rect, ElementEvent } from '../../';
+import { Renderer as CanvasRenderer } from '../../../../g-canvas';
 
 chai.use(chaiAlmost());
 chai.use(sinonChai);